+++ docs/CHANGELOG.md
@@ -43,7 +43,9 @@
     memory usage. It replaces `loop.FreeInput()`
 * Package `commandline`:
   * `AttachProgressBar` now displays a continuously updated table with metrics generated during training.
-    This only works in the commandline (not in notebooks).
+    This only works in the commandline (not in notebooks). 
+  * Asynchronous display of updates: it works better with very fast training loops or if running
+    over a slow terminal connection (network).
   * Added `CreateContextSettingsFlag` and `ParseContextSettings`.
 
 ## 0.8.0 - 2023/11/28